Replace router parsing regex with FSM.
The original algorithm for parsing route lines in the router used a
regular expression to group parts of the route line. However, this does
not work since it does not check for correctly matching parens. It is
impossible to use a regular expression to check for matching parens, so
this commit replaces the regular expression with a simple finite state
machine (FSM) that correctly validates routes.

Note that we should probably figure out how to test error conditions
like this so that we can more easily catch bugs like this.